Changed survival formula
now the relative effect of cost of help and benefit of group size is more intuitive. If Xsn=Xsh, equivalent size effect of help and group size in survival. X0 is independent, and therefore, we can increase of decrease the effect of help and group size independently (easier that the population doesn't go extinct)